Renamed base.py to test_base.py so that pytest rewrites the assertions for improved failure messages
