+++ tests/e2e/tests/test_smoketests.py
@@ -5,7 +5,7 @@
 import pytest
 
 from . import releng_utils as utils
-from .base import Base
+from .test_base import Base
 
 
 class TestSmokeTests(Base):
